Refactoring based on recommendations
