+++ pkg/controllers/aws_controllermanager.go
@@ -5,11 +5,16 @@ import (
 	"k8s.io/klog/v2"
 
 	cloudprovider "k8s.io/cloud-provider"
-	taggingcontroller "k8s.io/cloud-provider-aws/cmd/aws-cloud-controller-manager/controllers/tagging"
+	taggingcontroller "k8s.io/cloud-provider-aws/pkg/controllers/tagging"
 	"k8s.io/cloud-provider/app"
 	cloudcontrollerconfig "k8s.io/cloud-provider/app/config"
 	genericcontrollermanager "k8s.io/controller-manager/app"
 	"k8s.io/controller-manager/controller"
+)
+
+const (
+	TaggingControllerClientName = "tagging-controller"
+	TaggingControllerKey        = "tagging"
 )
 
 // BuildControllerInitializers is used to add new controllers built in this package to
@@ -19,12 +24,12 @@ func BuildControllerInitializers() map[string]app.ControllerInitFuncConstructor 
 
 	taggingControllerInitFuncConstrustor :=  app.ControllerInitFuncConstructor{
 		InitContext: app.ControllerInitContext{
-			ClientName: "tagging-controller",
+			ClientName: TaggingControllerClientName,
 		},
 		Constructor: startTaggingControllerWrapper,
 	}
 
-	controllerInitializers["tagging"] = taggingControllerInitFuncConstrustor
+	controllerInitializers[TaggingControllerKey] = taggingControllerInitFuncConstrustor
 
 	return controllerInitializers
 }
